+++ utils/patched_fetch.ts
